Don't delete default config
Neo4j config allows repeated keys for some types of
configuration; in other cases repeated keys simply override
prior keys.

The entrypoint currently deletes any prior key if a config
option is added as an env var; this means, notably, that all
the default JVM configuration is deleted if the user specifies
an additional JVM flag.

This fixes this as a bug - backported to 3.0 - and adds a regression
test.